CIRCSTORE-407 Move EventConsumerVerticleTest to another package
